[Bronze II] Title: 숫자, Time: 756 ms, Memory: 27760 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -5,21 +5,21 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
 
         if (a==b) {
             System.out.println(0);
             System.exit(0);
         }
 
-        int tmp1 = a;
-        int tmp2 = b;
+        long tmp1 = a;
+        long tmp2 = b;
         a = Math.min(tmp1, tmp2);
         b = Math.max(tmp1, tmp2);
 
         System.out.println(b-a-1);
-        for (int i = a+1; i < b; i++) {
+        for (long i = a+1; i < b; i++) {
             System.out.print(i+" ");
         }
     }
